requirements.txt
# file that contain the python packages used in the environment

geopandas 
matplotlib >=3.5.0
matplotlib-inline >= 0.1.2
numpy >= 1.20
pandas >=1.3.4
xarray >=0.19.0
scikit-learn
IPython
pytorch 
scipy
tqdm
scipy
torchvision
#fastai
rasterio
opencv
patchify
albumentations
